# Practice2/src/main/java/es/mastercloud/practice2/models/dto/InBookDTO.java
package es.mastercloud.practice2.models.dto;

import java.util.List;

public record InBookDTO (
        String title,
        String synopsis,
        String author,
        String publisher,
        String publishYear,
        List<BookReviewDTO> reviews
){}